changed all uint64_t to int64_t, some cleanup

+++ BinData.h
@@ -52,7 +52,7 @@ class BinData : public BinDataBase {
   BinData& operator= ( BinData& ) = delete;
   BinData& operator= ( BinData&& ) = delete;
 
-  int getMaxElements ( QString filename, int sigIndex, uint64_t& max );
+  int getMaxElements ( QString filename, int sigIndex, int64_t& max );
 
   int genLineSeries ( QString filename,
                      int sigIndex,
@@ -62,17 +62,17 @@ class BinData : public BinDataBase {
                      double startTimeInSec,
                      double endTimeInSec,
                      double dataTimeIncrementInSec,
-                     uint64_t& numPts,
+                     int64_t& numPts,
                      QtCharts::QLineSeries& qls,
                      double& miny,
                      double& maxy,
-                     uint64_t maxFft,
-                     uint64_t& numFft,
+                     int64_t maxFft,
+                     int64_t& numFft,
                      fftw_complex *array  );
 
-  int readTraceData( std::filebuf& fb, int *buf, int readSizeInbytes );
+  int64_t readTraceData( std::filebuf& fb, int *buf, int64_t readSizeInbytes );
 
-  int getMaxElements2 ( QString filename, int sigIndex, uint64_t& max );
+  int getMaxElements2 ( QString filename, int sigIndex, int64_t& max );
 
   int genLineSeries2 ( QString filename,
                      int sigIndex,
@@ -82,18 +82,18 @@ class BinData : public BinDataBase {
                      double startTimeInSec,
                      double endTimeInSec,
                      double dataTimeIncrementInSec,
-                     uint64_t& numPts,
+                     int64_t& numPts,
                      QtCharts::QLineSeries& qls,
                      double& miny,
                      double& maxy,
-                     uint64_t maxFft,
-                     uint64_t& numFft,
+                     int64_t maxFft,
+                     int64_t& numFft,
                      fftw_complex *array  );
 
-  int readTraceData2 ( std::filebuf& fb, int *buf, int readSizeInbytes );
+  int64_t readTraceData2 ( std::filebuf& fb, int *buf, int64_t readSizeInbytes );
 
   void updateLineSeries (
-    int readOpCount,
+    int64_t readOpCount,
     QPointF *pts,
     double slope,
     double intercept,
@@ -102,7 +102,7 @@ class BinData : public BinDataBase {
     double startTimeInSec,
     double endTimeInSec,
     double dataTimeIncrementInSec,
-    int numBytesToProcess,
+    int64_t numBytesToProcess,
     int *buf,
     LineSeriesBuilderBase *lsb,
     double& miny,
